Updated types on methods and fields

+++ src/Interfaces/Item/ItemPropertiesBarrelInterface.php
@@ -6,14 +6,14 @@ namespace App\Interfaces\Item;
 
 interface ItemPropertiesBarrelInterface
 {
-    public function getErgonomics(): int;
-    public function setErgonomics(int $ergonomics): ItemPropertiesBarrelInterface;
-    public function getRecoilModifier(): int;
-    public function setRecoilModifier(int $recoilModifier): ItemPropertiesBarrelInterface;
-    public function getCenterOfImpact(): int;
-    public function setCenterOfImpact(int $centerOfImpact): ItemPropertiesBarrelInterface;
-    public function getDeviationCurve(): int;
-    public function setDeviationCurve(int $deviationCurve): ItemPropertiesBarrelInterface;
-    public function getDeviationMax(): int;
-    public function setDeviationMax(int $deviationMax): ItemPropertiesBarrelInterface;
+    public function getErgonomics(): float;
+    public function setErgonomics(float $ergonomics): ItemPropertiesBarrelInterface;
+    public function getRecoilModifier(): float;
+    public function setRecoilModifier(float $recoilModifier): ItemPropertiesBarrelInterface;
+    public function getCenterOfImpact(): float;
+    public function setCenterOfImpact(float $centerOfImpact): ItemPropertiesBarrelInterface;
+    public function getDeviationCurve(): float;
+    public function setDeviationCurve(float $deviationCurve): ItemPropertiesBarrelInterface;
+    public function getDeviationMax(): float;
+    public function setDeviationMax(float $deviationMax): ItemPropertiesBarrelInterface;
 }
